+++ Sora/Configuration.swift
@@ -161,7 +161,6 @@ public struct Configuration {
     // MARK: - イベントハンドラ
 
     /// WebSocket チャネルに関するイベントハンドラ
-    @available(*, unavailable, message: "廃止されました。 mediaChannelHandlers を利用してください。")
     public var webSocketChannelHandlers = WebSocketChannelHandlers()
 
     /// シグナリングチャネルに関するイベントハンドラ